+++ tests/test_composite.py
