+++ tests/test_reads.py
@@ -1,14 +1,10 @@
-import math
 import random
 
 import aiohttp
-from morecantile.models import TileMatrixSet
 import mercantile
 import numpy as np
 import pytest
-from PIL import Image
 import rasterio
-from rasterio.crs import CRS
 from rasterio.warp import transform_bounds
 from rasterio.windows import Window
 from rio_tiler.io.cogeo import COGReader as cogeo_reader
@@ -16,70 +12,10 @@ from rio_tiler import utils as rio_tiler_utils
 from rio_tiler.mercator import get_zooms
 from shapely.geometry import Polygon
 
-from aiocogeo import config, COGReader
-from aiocogeo.ifd import IFD
-from aiocogeo.tag import BaseTag
-from aiocogeo.errors import InvalidTiffError, TileNotFoundError
-from aiocogeo.constants import MaskFlags
+from aiocogeo import config
+from aiocogeo.errors import TileNotFoundError
 
 from .conftest import TEST_DATA
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("infile", TEST_DATA)
-async def test_cog_metadata(infile, create_cog_reader):
-    async with create_cog_reader(infile) as cog:
-        with rasterio.open(infile) as ds:
-            rio_profile = ds.profile
-            cog_profile = cog.profile
-
-            # Don't compare photometric, rasterio seems to not always report color interp
-            cog_profile.pop("photometric", None)
-            rio_profile.pop("photometric", None)
-
-            assert [member.value for member in ds.colorinterp] == [member.value for member in cog.color_interp]
-            assert rio_profile == cog_profile
-            assert ds.overviews(1) == cog.overviews
-
-            rio_tags = ds.tags()
-            gdal_metadata = cog.gdal_metadata
-
-            for (k,v) in rio_tags.items():
-                if k in ("TIFFTAG_XRESOLUTION", "TIFFTAG_YRESOLUTION", "TIFFTAG_RESOLUTIONUNIT"):
-                    continue
-                assert str(gdal_metadata[k]) == v
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("infile", TEST_DATA)
-async def test_cog_metadata_overviews(infile, create_cog_reader):
-    async with create_cog_reader(infile) as cog:
-        for idx, ifd in enumerate(cog.ifds):
-            width = ifd.ImageWidth.value
-            height = ifd.ImageHeight.value
-            try:
-                # Test decimation of 2
-                next_ifd = cog.ifds[idx + 1]
-                next_width = next_ifd.ImageWidth.value
-                next_height = next_ifd.ImageHeight.value
-                assert pytest.approx(width / next_width, 5) == 2.0
-                assert pytest.approx(height / next_height, 5) == 2.0
-
-                # Test number of tiles
-                tile_count = ifd.tile_count[0] * ifd.tile_count[1]
-                next_tile_count = next_ifd.tile_count[0] * next_ifd.tile_count[1]
-                assert (
-                    pytest.approx(
-                        (
-                            max(tile_count, next_tile_count)
-                            / min(tile_count, next_tile_count)
-                        ),
-                        3,
-                    )
-                    == 4.0
-                )
-            except IndexError:
-                pass
 
 
 @pytest.mark.asyncio
@@ -403,41 +339,6 @@ async def test_boundless_get_tile(create_cog_reader, infile, monkeypatch):
 
 
 @pytest.mark.asyncio
-async def test_cog_has_alpha_band(create_cog_reader):
-    async with create_cog_reader("https://async-cog-reader-test-data.s3.amazonaws.com/cog_alpha_band.tif") as cog:
-        assert cog.has_alpha
-
-    async with create_cog_reader(TEST_DATA[0]) as cog:
-        assert not cog.has_alpha
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("infile,expected", zip(TEST_DATA, [
-    [[MaskFlags.all_valid], [MaskFlags.all_valid], [MaskFlags.all_valid]],
-    [[MaskFlags.all_valid], [MaskFlags.all_valid], [MaskFlags.all_valid]],
-    [[MaskFlags.all_valid], [MaskFlags.all_valid], [MaskFlags.all_valid]],
-    [[MaskFlags.all_valid], [MaskFlags.all_valid], [MaskFlags.all_valid]],
-    [[MaskFlags.all_valid], [MaskFlags.all_valid], [MaskFlags.all_valid]],
-    [[MaskFlags.nodata], [MaskFlags.nodata], [MaskFlags.nodata]],
-    [[MaskFlags.nodata], [MaskFlags.nodata], [MaskFlags.nodata]],
-    [[MaskFlags.per_dataset], [MaskFlags.per_dataset], [MaskFlags.per_dataset]],
-    [[MaskFlags.nodata], [MaskFlags.nodata], [MaskFlags.nodata]],
-    [
-        [MaskFlags.per_dataset, MaskFlags.alpha],
-        [MaskFlags.per_dataset, MaskFlags.alpha],
-        [MaskFlags.per_dataset, MaskFlags.alpha],
-        [MaskFlags.all_valid]
-    ],
-    [[MaskFlags.per_dataset]],
-    [[MaskFlags.all_valid], [MaskFlags.all_valid], [MaskFlags.all_valid]],
-
-]))
-async def test_cog_mask_flags(create_cog_reader, infile, expected):
-    async with create_cog_reader(infile) as cog:
-        mask_flags = cog.mask_flags
-    assert expected == mask_flags
-
-@pytest.mark.asyncio
 @pytest.mark.parametrize("infile", TEST_DATA)
 async def test_read_not_in_bounds(create_cog_reader, infile):
     tile = mercantile.Tile(x=0, y=0, z=25)
@@ -448,17 +349,6 @@ async def test_read_not_in_bounds(create_cog_reader, infile):
             bounds = transform_bounds("EPSG:3857", f"EPSG:{cog.epsg}", *bounds)
         with pytest.raises(TileNotFoundError):
             await cog.read(bounds=bounds, shape=(256, 256))
-
-
-@pytest.mark.asyncio
-async def test_cog_palette(create_cog_reader):
-    infile = "https://async-cog-reader-test-data.s3.amazonaws.com/cog_cmap.tif"
-    async with create_cog_reader(infile) as cog:
-        with rasterio.open(infile) as ds:
-            cog_interp = cog.color_interp
-            rio_interp = ds.colorinterp
-            assert cog_interp[0].value == rio_interp[0].value
-            assert cog.colormap == ds.colormap(1)
 
 
 @pytest.mark.asyncio
@@ -476,113 +366,6 @@ async def test_cog_get_overview_level(create_cog_reader, width, height):
             # Our index for source data is 0 while rio tiler uses -1
             expected_ovr = 0 if expected_ovr == -1 else expected_ovr
             assert ovr == expected_ovr
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("infile", TEST_DATA)
-async def test_cog_tile_matrix_set(infile, create_cog_reader):
-    async with create_cog_reader(infile) as cog:
-        tile_matrix_set = cog.create_tile_matrix_set()
-        TileMatrixSet(**tile_matrix_set)
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("infile", [TEST_DATA[0]])
-async def test_cog_metadata_iter(infile, create_cog_reader):
-    async with create_cog_reader(infile) as cog:
-        for ifd in cog:
-            assert isinstance(ifd, IFD)
-            for tag in ifd:
-                assert isinstance(tag, BaseTag)
-
-#
-@pytest.mark.asyncio
-async def test_block_cache_enabled(create_cog_reader, monkeypatch):
-    # Cache is disabled for tests
-    monkeypatch.setattr(config, "ENABLE_BLOCK_CACHE", True)
-    infile = "https://async-cog-reader-test-data.s3.amazonaws.com/lzw_cog.tif"
-    async with create_cog_reader(infile) as cog:
-        await cog.get_tile(0, 0, 0)
-
-    async with create_cog_reader(infile) as cog:
-        await cog.get_tile(0, 0, 0)
-        # Confirm all requests are cached
-        assert cog.requests["count"] == 18
-
-
-@pytest.mark.asyncio
-async def test_block_cache_disabled(create_cog_reader):
-    infile = "https://async-cog-reader-test-data.s3.amazonaws.com/lzw_cog.tif"
-    async with create_cog_reader(infile) as cog:
-        await cog.get_tile(0, 0, 0)
-        request_count = cog.requests["count"]
-
-        await cog.get_tile(0, 0, 0)
-        assert cog.requests["count"] == request_count + 1
-
-
-@pytest.mark.asyncio
-async def test_header_cache_enabled(create_cog_reader, monkeypatch):
-    # Cache is disabled for tests
-    monkeypatch.setattr(config, "ENABLE_HEADER_CACHE", True)
-    infile = "https://async-cog-reader-test-data.s3.amazonaws.com/webp_cog.tif"
-    async with create_cog_reader(infile) as cog:
-        assert cog.requests["count"] == 20
-
-    async with create_cog_reader(infile) as cog:
-        assert cog.requests["count"] == 2
-
-    async with create_cog_reader(infile) as cog:
-        await cog.get_tile(0, 0, 0)
-        assert cog.requests["count"] == 3
-
-
-@pytest.mark.asyncio
-async def test_header_cache_disabled(create_cog_reader):
-    infile = "https://async-cog-reader-test-data.s3.amazonaws.com/webp_cog.tif"
-    async with create_cog_reader(infile) as cog:
-        assert cog.requests["count"] == 20
-
-    async with create_cog_reader(infile) as cog:
-        assert cog.requests["count"] == 20
-
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("infile", TEST_DATA)
-async def test_cog_request_metadata(create_cog_reader, infile):
-    async with create_cog_reader(infile) as cog:
-        request_metadata = cog.requests
-
-    assert len(request_metadata["ranges"]) == request_metadata["count"]
-    assert (
-        sum([end - start + 1 for (start, end) in request_metadata["ranges"]])
-        == request_metadata["byte_count"]
-    )
-
-
-@pytest.mark.asyncio
-async def test_cog_not_a_tiff(create_cog_reader):
-    infile = "https://async-cog-reader-test-data.s3.amazonaws.com/not_a_tiff.png"
-    with pytest.raises(InvalidTiffError):
-        async with create_cog_reader(infile) as cog:
-            ...
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    "infile",
-    [
-        "/local/file/does/not/exist",
-        "https://cogsarecool.com/cog.tif",  # Invalid host
-        "https://async-cog-reader-test-data.s3.amazonaws.com/file-does-not-exist.tif",  # valid host invalid path
-        "s3://nobucket/badkey.tif",
-    ],
-)
-async def test_file_not_found(create_cog_reader, infile):
-    with pytest.raises(FileNotFoundError):
-        async with create_cog_reader(infile) as cog:
-            ...
 
 
 @pytest.mark.asyncio
